requirements.txt
Django
djangorestframework
requests
# venv/bin/pip install --no-cache-dir -i https://pypi.douban.com/simple -r requirements.txt